# Run the update status of installed and transitive packages in background