<!-- DO NOT MODIFY OR DELETE THIS TEMPLATE. IT IS USED IN AUTOMATION. -->

# Bug

<!-- If this is an engineering change or test change only, you do not need an issue. -->
<!-- Find or create an issue in NuGet/Home and paste the full url. -->
<!-- At the maintainers discretion, multiple changes may apply to a single issue, but only when the PRs are all created within a short period of time. -->
Fixes: https://github.com/NuGet/Home/issues/14256

## Description
This change will yield the packages list to the caller before it finishes running the expensive operations to populate extra metadata for the Installed packages. Previously, the expensive operations were being synchronously ran and blocking the caller thread until finished, causing the packages list to not be returned for display until the versions and metadata was also fetched, along with some other computations that happened beforehand.

Taking the duration measurement (in milliseconds) for the PMUIRefresh when VS.NuGet.RefreshSource is PackageManagerLoaded for the after and before cases, using OrchardCore as the same solution, and loading the Solution PM UI produced these results:

Run | Before Duration | After Duration
-- | -- | --
Run 1 | 65332.9408 | 11206.2541
Run 2 | 16534.3167 | 13645.67
Run 3 | 17677.6345 | 15105.5598
Run 4 | 19484.0274 | 13327.0325
Run 5 | 17657.166 | 10660.1747

The average runtime for the Before durations is 27337.22 ms and for the After durations is 12788.94 ms. The percentage increase in performance from the After duration compared to Before is **53.22%**. Removing the outliers from the Before and After (the highest and lowest), the increase in performance from the After duration compared to Before is **30.35%**.

These metrics are subject to variation depending on the user's network and other computer factors. However, the slower the network connection, the greater the improvement is expected to be.

The behavior change is that the packages list will load faster on the Installed tab and the Versions, Metadata (including vulnerability information and deprecations) will be computed in the background and updated as soon as available. This will drastically reduce the time until a user can interact with the packages list in the PM UI.

## PR Checklist

- [x] Meaningful title, helpful description and a linked NuGet/Home issue
- [ ] ~~Added tests~~
- [ ] ~~Link to an issue or pull request to update docs if this PR changes settings, environment variables, new feature, etc.~~