@jebriede 2025-04-29
@Nigusu-Allehu answers below.

> LGTM but I have some concerns
> 
> * When we display a package and later update its metadata, could this cause the UI to flicker .Especially in cases where metadata loading takes longer than usual?

No, it does not flicker. It loads the package then updates the extra data as it becomes available. Essentially, it "fills in" when available.

> * if a user switches tabs multiple times, is there a risk that we could schedule multiple metadata update tasks for the same package, leading to redundant work?

This change does not introduce an added risk for scheduling multiple update tasks. The UI was still available before this change and switching tabs would also cause a request for metadata to be queried. There are however extra mechanisms to prevent a lot of the metadata from being queried twice. The new package model has protection against this, and the existing code has caching. We can always add cancellation tokens in the future if we want to explicitly cancel the operation if requested more than once but I'll defer that to a future change to keep these changed isolated.